C2 - Guessing
Removed a few files due to changing the proposed logic and structure. Added Guess.jsx which displays the values of an entered guess with some temporary styling. Also added Search.jsx, which allows users to input aa guess via a search bar. Linked the logic for these in App.jsx, alllowing for guesses to be made.